server/requirements.txt
fastapi
uvicorn
easyocr
torch
transformers
pillow
python-multipart
numpy
google-generativeai
python-dotenv
